:sparkles: :: icon button 코드 이동 및 button component 추가
