+++ app/src/main/java/com/caramm/textextractor/core/designsystem/TextExtractorIcon.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -24,24 +23,6 @@ fun TextExtractorIcon(
         modifier = modifier,
         tint = tint
     )
-}
-
-@Composable
-fun TextExtractorIconButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    imageVector: ImageVector = Icons.Default.Add,
-    tint: Color = TextExtractorTheme.colors.text
-) {
-    IconButton(
-        onClick = onClick,
-        modifier = modifier
-    ) {
-        TextExtractorIcon(
-            imageVector = imageVector,
-            tint = tint
-        )
-    }
 }
 
 @Preview
